-- Added CUDA-accelerated plane registration through a new GpuRegistrationBackend, selected per job by the `--register-device` option of `cindra run` and the `gpu_devices` argument of the execute MCP tools, with the host CPU remaining the default.
-- Added CUDA device discovery and runtime verification in `cindra.orchestration.gpu`, surfaced through the `cindra gpu` command and `check_gpu_runtime_tool`, so a batch gates on a readiness flag rather than on a per-job tracker failure.
-- Added a device-backed registration resource class that caps its concurrency at the device count and charges each job the 2-core `REGISTRATION_GPU_WORKERS` default, since a device-backed job keeps only its reference, crop, and median-filter work on the host.
-- Added dispatch-time worker widening for the elastic stages, dividing the free cores among the classes holding queued work so a draining queue widens registration, discovery, and extraction jobs toward their published ceilings.
-- Added `pipeline_type` to every `execute_full_pipeline_tool` outcome, so a caller can attribute any response, including an argument rejection, to the pipeline it requested.
-- Optimized the cross-recording B-spline fit and grid sampling.
-- Corrected 55 documentation claims that disagreed with the code, and applied 210 style corrections across the README, the plugin skills, and the library docstrings.
-- Fixed various repository hygiene issues, covering the tracked local settings file, the abbreviated bidirectional phase correction module name, and two subprocess checks ruff 0.16.5 began reporting.

+++ src/cindra/io/combine.py
@@ -456,8 +456,8 @@ def _compute_plane_offsets(
         if roi_number < plane_number:
             z_plane_number = plane_number // roi_number
 
-            heights_array = np.array([context.runtime.io.frame_height for context in plane_contexts])
-            widths_array = np.array([context.runtime.io.frame_width for context in plane_contexts])
+            heights_array = np.array([context.runtime.io.frame_height for context in plane_contexts], dtype=np.uint16)
+            widths_array = np.array([context.runtime.io.frame_width for context in plane_contexts], dtype=np.uint16)
 
             # Calculates the tile size as the bounding box that contains all ROIs at their MROI positions.
             maximum_height = (y_displacement + heights_array).max()
